% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_models.R
\name{ll.models}
\alias{ll.models}
\title{MLE estimation for linear and logistic regression models}
\usage{
ll.models(y, X, model.type)
}
\arguments{
\item{y}{Vector of the outcome variable or dependent variable}

\item{X}{Matrix of the independent variables.}

\item{model.type}{Specifies whether the function calculates MLE parameter estimates and associated standard errors for
logistic (link logit) or linear models}
}
\value{
Returns a matrix of parameter estimates and associated standard errors
}
\description{
MLE estimation for linear and logistic regression models
}
\examples{
model.linear <- ll.models(y=y, X=X, model.type='linear')
model.logistic <- ll.models(y=y, X=X, model.type='logistic')

}
